feat(#1421): move class signature to DirectivesClass
