+++ src/test/java/org/eolang/jeo/representation/directives/DirectivesClassPropertiesTest.java
@@ -30,7 +30,6 @@ final class DirectivesClassPropertiesTest {
                         new DirectivesClassProperties(
                             1,
                             "org/eolang/SomeClass",
-                            "java/lang/Object",
                             "org/eolang/SomeInterface"
                         )
                     ).up()
@@ -38,59 +37,9 @@ final class DirectivesClassPropertiesTest {
             XhtmlMatchers.hasXPaths(
                 "/o/o[contains(@base,'number') and contains(@name,'version')]",
                 "/o/o[contains(@base,'number') and contains(@name,'access')]",
-                "/o/o[contains(@base,'string') and contains(@name,'signature')]",
                 "/o/o[contains(@base,'string') and contains(@name,'supername')]",
                 new JeoBaseXpath("./o/o", "seq.of1").toXpath(),
                 "/o/o[contains(@name,'interfaces')]"
-            )
-        );
-    }
-
-    @Test
-    void createsDirectivesWithMandatorySignature() throws ImpossibleModificationException {
-        MatcherAssert.assertThat(
-            "Can't create proper xml with mandatory signature",
-            new Xembler(
-                new Directives()
-                    .add("o")
-                    .append(
-                        new DirectivesClassProperties(
-                            1,
-                            null,
-                            "java/lang/Object",
-                            "org/eolang/SomeInterface"
-                        )
-                    ).up()
-            ).xml(),
-            XhtmlMatchers.hasXPaths(
-                "//o[contains(@name,'signature')]"
-            )
-        );
-    }
-
-    /**
-     * This test was added to ensure that when a null signature is passed,
-     * it is converted to an empty string in the XML output.
-     * See more details in
-     * <a href="https://github.com/objectionary/jeo-maven-plugin/issues/1246">#1246</a>
-     * @throws ImpossibleModificationException in case of XML modification failure.
-     */
-    @Test
-    void convertsNullSignatureToEmptyString() throws ImpossibleModificationException {
-        MatcherAssert.assertThat(
-            "Signature should be empty string when null is passed",
-            new Xembler(
-                new Directives().add("o").append(
-                    new DirectivesClassProperties(
-                        1,
-                        null,
-                        "java/lang/Object",
-                        "org/eolang/SomeInterface"
-                    )
-                ).up()
-            ).xml(),
-            XhtmlMatchers.hasXPath(
-                "//o[@name='signature']/o[contains(@base,'bytes')]/o[text()='--']"
             )
         );
     }
